自定义异常处理 枚举状态码ç及信息 回复问题

+++ src/main/java/com/yh_simon/community/controller/PublishController.java
@@ -24,7 +24,7 @@ public class PublishController {
     QuestionService questionService;
 
     @GetMapping("/publish/{id}")
-    public String publish(@PathVariable("id")Integer id,Model model){
+    public String publish(@PathVariable("id")Long id,Model model){
         Question question=questionMapper.getById(id);
         model.addAttribute("title", question.getTitle());
         model.addAttribute("description", question.getDescription());
@@ -42,7 +42,7 @@ public class PublishController {
     public String doPublish(@RequestParam(value = "title", required = false) String title,
                             @RequestParam(value = "description", required = false) String description,
                             @RequestParam(value = "tag", required = false) String tag,
-                            @RequestParam(value = "id",required = false) Integer id,
+                            @RequestParam(value = "id",required = false) Long id,
                             HttpServletRequest request, Model model) {
